requirements.txt
numpy
aiohttp
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
neuralgcm
tenacity
